fallback image valid step
